Added add to playlist, delete playlist, delete from playlist

+++ app/src/main/java/fr/epf/min2/countries_app/ui/item/ActivityItemCountry.kt
@@ -29,7 +29,7 @@ class ActivityItemCountry : AppCompatActivity() {
         val countryCurrency = intent.getStringExtra("COUNTRY_CURRENCY")
         val countryLanguage = intent.getStringExtra("COUNTRY_LANGUAGE")
         val countryDemonym = intent.getStringExtra("COUNTRY_DEMONYM")
-        val countryPopulation = intent.getIntExtra("COUNTRY_POPULATION",0,)
+        val countryPopulation = intent.getLongExtra("COUNTRY_POPULATION",0L,)
         val countryDrivesOn = intent.getStringExtra("COUNTRY_DRIVES_ON")
         val countryFlagUrl = intent.getStringExtra("COUNTRY_FLAG_URL")
         val countryLat = intent.getDoubleExtra("COUNTRY_LAT", 0.0)
